Update front end to function with localhost.
This set of changes replaces the function for ImgMediaCard with a
class, which caused me to need to redefine some things.

This change was made so I could hook into onComponentDidMount, so the
image that starts on the page could be dynamically loaded in the right
place in the lifecycle.

This also changes the functionality such that the app fetches an image
from my localhost api. Imagining you have a localhost api that can
produce the images you want to load, this all functions now.